requirements.txt
aiohappyeyeballs>=2.4.0
aiohttp>=3.10.6
aiosignal>=1.3.1
annotated-types>=0.7.0
anyio>=4.6.0
attrs>=24.2.0
certifi>=2024.8.30
charset-normalizer>=3.3.2
colorama>=0.4.6
dataclasses-json>=0.6.7
distro>=1.9.0
frozenlist>=1.4.1
greenlet>=3.1.1
h11>=0.14.0
httpcore>=1.0.5
httpx>=0.27.2
idna>=3.10
jiter>=0.5.0
jsonpatch>=1.33
jsonpointer>=3.0.0
langchain>=0.3.1
langchain-community>=0.3.1
langchain-core>=0.3.6
langchain-openai>=0.2.1
langchain-text-splitters>=0.3.0
langgraph>=0.2.28
langgraph-checkpoint>=1.0.11
langsmith>=0.1.128
marshmallow>=3.22.0
msgpack>=1.1.0
multidict>=6.1.0
mypy-extensions>=1.0.0
numpy>=1.26.4
openai>=1.50.0
orjson>=3.10.7
packaging>=24.1
pydantic>=2.9.2
pydantic-settings>=2.5.2
pydantic_core>=2.23.4
python-dotenv>=1.0.1
PyYAML>=6.0.2
regex>=2024.9.11
requests>=2.32.3
sniffio>=1.3.1
SQLAlchemy>=2.0.35
tenacity>=8.5.0
tiktoken>=0.7.0
tqdm>=4.66.5
typing-inspect>=0.9.0
typing_extensions>=4.12.2
urllib3>=2.2.3
yarl>=1.12.1
tavily-python
langgraph_sdk 
